chore(EMS-4116): policy - check your answers - align, simplify e2e tests
