+++ .github/workflows/test.yml
@@ -679,7 +679,9 @@ jobs:
             'policy/change-your-answers/other-company-details/**/*.spec.js',
             'policy/change-your-answers/pre-credit-period/**/*.spec.js',
             'policy/change-your-answers/single-policy/**/*.spec.js',
-            'policy/check-your-answers/**/*.spec.js',
+            'policy/check-your-answers/*.spec.js',
+            'policy/check-your-answers/summary-list/single-policy/**/*.spec.js',
+            'policy/check-your-answers/summary-list/multiple-policy/**/*.spec.js',
             'policy/different-name-on-policy/**/*.spec.js',
             'policy/loss-payee/**/*.spec.js',
             'policy/loss-payee-details/**/*.spec.js',
