+++ e2e-tests/insurance/cypress/e2e/journeys/policy/check-your-answers/summary-list/multiple-policy/check-your-answers-multiple-loss-payee-international.spec.js
@@ -1,7 +1,7 @@
-import { FIELD_VALUES } from '../../../../../../../constants';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
-import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance/policy';
-import checkSummaryList from '../../../../../../../commands/insurance/check-policy-summary-list';
+import { FIELD_VALUES } from '../../../../../../../../constants';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
+import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
 
 const { ROOT: INSURANCE_ROOT, POLICY } = INSURANCE_ROUTES;
 
@@ -15,7 +15,7 @@ const {
 
 const baseUrl = Cypress.config('baseUrl');
 
-context('Insurance - Policy - Check your answers - Summary list - Multiple contract policy - Using loss payee based in the UK', () => {
+context('Insurance - Policy - Check your answers - Summary list - Multiple contract policy - Loss payee based internationally', () => {
   let referenceNumber;
   let url;
 
@@ -26,7 +26,7 @@ context('Insurance - Policy - Check your answers - Summary list - Multiple contr
       cy.completePolicySection({
         policyType: FIELD_VALUES.POLICY_TYPE.MULTIPLE,
         isAppointingLossPayee: true,
-        lossPayeeIsLocatedInUK: true,
+        lossPayeeIsLocatedInUK: false,
       });
 
       url = `${baseUrl}${INSURANCE_ROOT}/${referenceNumber}${POLICY.CHECK_YOUR_ANSWERS}`;
@@ -43,10 +43,6 @@ context('Insurance - Policy - Check your answers - Summary list - Multiple contr
     cy.deleteApplication(referenceNumber);
   });
 
-  it('should render generic policy summary list rows', () => {
-    cy.assertGenericMultiplePolicySummaryListRows();
-  });
-
   it(`should render a ${LOSS_PAYEE_IS_APPOINTED} summary list row`, () => {
     checkSummaryList[LOSS_PAYEE_IS_APPOINTED]({ isAppointingLossPayee: true });
   });
@@ -56,22 +52,22 @@ context('Insurance - Policy - Check your answers - Summary list - Multiple contr
   });
 
   it(`should render a ${FINANCIAL_ADDRESS} summary list row`, () => {
-    checkSummaryList.LOSS_PAYEE[FINANCIAL_ADDRESS]({ shouldRender: true, isUk: true });
+    checkSummaryList.LOSS_PAYEE[FINANCIAL_ADDRESS]({ shouldRender: true, isInternational: true });
   });
 
-  it(`should render a ${SORT_CODE} summary list row`, () => {
-    checkSummaryList.LOSS_PAYEE[SORT_CODE]({ shouldRender: true });
+  it(`should render a ${BIC_SWIFT_CODE} summary list row`, () => {
+    checkSummaryList.LOSS_PAYEE[BIC_SWIFT_CODE]({ shouldRender: true });
   });
 
-  it(`should render a ${ACCOUNT_NUMBER} summary list row`, () => {
-    checkSummaryList.LOSS_PAYEE[ACCOUNT_NUMBER]({ shouldRender: true });
+  it(`should render an ${IBAN} summary list row`, () => {
+    checkSummaryList.LOSS_PAYEE[IBAN]({ shouldRender: true });
   });
 
-  it(`should NOT render a ${BIC_SWIFT_CODE} summary list row`, () => {
-    checkSummaryList.LOSS_PAYEE[BIC_SWIFT_CODE]({ shouldRender: false });
+  it(`should NOT render a ${SORT_CODE} summary list row`, () => {
+    checkSummaryList.LOSS_PAYEE[SORT_CODE]({ shouldRender: false });
   });
 
-  it(`should NOT render an ${IBAN} summary list row`, () => {
-    checkSummaryList.LOSS_PAYEE[IBAN]({ shouldRender: false });
+  it(`should NOT render a ${ACCOUNT_NUMBER} summary list row`, () => {
+    checkSummaryList.LOSS_PAYEE[ACCOUNT_NUMBER]({ shouldRender: false });
   });
 });
