+++ e2e-tests/insurance/cypress/e2e/journeys/policy/check-your-answers/summary-list/single-policy/check-your-answers-single-alternative-currency.spec.js
@@ -1,7 +1,7 @@
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
-import { INSURANCE_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance';
-import checkSummaryList from '../../../../../../../commands/insurance/check-policy-summary-list';
-import { NON_STANDARD_CURRENCY_CODE, NON_STANDARD_CURRENCY_NAME } from '../../../../../../../fixtures/currencies';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { INSURANCE_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance';
+import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
+import { NON_STANDARD_CURRENCY_CODE, NON_STANDARD_CURRENCY_NAME } from '../../../../../../../../fixtures/currencies';
 
 const { ROOT: INSURANCE_ROOT, POLICY } = INSURANCE_ROUTES;
