+++ e2e-tests/insurance/cypress/e2e/journeys/policy/check-your-answers/summary-list/single-policy/check-your-answers-single-broker-based-in-uk-summary-list.spec.js
@@ -1,0 +1,54 @@
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
+import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
+
+const { ROOT: INSURANCE_ROOT, POLICY } = INSURANCE_ROUTES;
+
+const {
+  USING_BROKER,
+  BROKER_DETAILS: { NAME, EMAIL },
+  BROKER_ADDRESSES: { SELECT_THE_ADDRESS },
+} = POLICY_FIELD_IDS;
+
+const baseUrl = Cypress.config('baseUrl');
+
+context('Insurance - Policy - Check your answers - Summary list - Single contract policy - Broker - based in UK - Summary List', () => {
+  let url;
+  let referenceNumber;
+
+  before(() => {
+    cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
+      referenceNumber = refNumber;
+
+      cy.completePolicySection({ usingBroker: true, brokerIsBasedInUk: true });
+
+      url = `${baseUrl}${INSURANCE_ROOT}/${referenceNumber}${POLICY.CHECK_YOUR_ANSWERS}`;
+    });
+  });
+
+  beforeEach(() => {
+    cy.saveSession();
+
+    cy.navigateToUrl(url);
+  });
+
+  after(() => {
+    cy.deleteApplication(referenceNumber);
+  });
+
+  it(`should render a ${USING_BROKER} summary list row`, () => {
+    checkSummaryList[USING_BROKER]({ usingBroker: true });
+  });
+
+  it(`should render a ${NAME} summary list row`, () => {
+    checkSummaryList.BROKER[NAME]({});
+  });
+
+  it(`should render a ${SELECT_THE_ADDRESS} summary list row`, () => {
+    checkSummaryList.BROKER[SELECT_THE_ADDRESS]();
+  });
+
+  it(`should render a ${EMAIL} summary list row`, () => {
+    checkSummaryList.BROKER[EMAIL]();
+  });
+});
