+++ e2e-tests/insurance/cypress/e2e/journeys/policy/check-your-answers/summary-list/single-policy/check-your-answers-single-need-pre-credit-period.spec.js
@@ -1,16 +1,10 @@
-import checkSummaryList from '../../../../../../../commands/insurance/check-policy-summary-list';
-import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../constants/field-ids/insurance/policy';
-import { INSURANCE_ROUTES } from '../../../../../../../constants/routes/insurance';
+import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
 
 const { ROOT: INSURANCE_ROOT, POLICY } = INSURANCE_ROUTES;
 
-const {
-  NEED_PRE_CREDIT_PERIOD,
-  CREDIT_PERIOD_WITH_BUYER,
-  NAME_ON_POLICY: { NAME },
-  USING_BROKER,
-  LOSS_PAYEE: { IS_APPOINTED: LOSS_PAYEE_IS_APPOINTED },
-} = POLICY_FIELD_IDS;
+const { NEED_PRE_CREDIT_PERIOD, CREDIT_PERIOD_WITH_BUYER } = POLICY_FIELD_IDS;
 
 const baseUrl = Cypress.config('baseUrl');
 
@@ -38,27 +32,11 @@ context('Insurance - Policy - Check your answers - Summary list - Single contrac
     cy.deleteApplication(referenceNumber);
   });
 
-  it('should render generic policy summary list rows', () => {
-    cy.assertGenericSinglePolicySummaryListRows();
-  });
-
   it(`should render a ${NEED_PRE_CREDIT_PERIOD} summary list row`, () => {
     checkSummaryList[NEED_PRE_CREDIT_PERIOD]({ needPreCreditPeriod: true });
   });
 
   it(`should render a ${CREDIT_PERIOD_WITH_BUYER} summary list row`, () => {
     checkSummaryList[CREDIT_PERIOD_WITH_BUYER]({ shouldRender: true });
-  });
-
-  it(`should render a ${NAME} summary list row`, () => {
-    checkSummaryList[NAME]({});
-  });
-
-  it(`should render a ${USING_BROKER} summary list row`, () => {
-    checkSummaryList[USING_BROKER]({ usingBroker: false });
-  });
-
-  it(`should render a ${LOSS_PAYEE_IS_APPOINTED} summary list row`, () => {
-    checkSummaryList[LOSS_PAYEE_IS_APPOINTED]({ isAppointingLossPayee: false });
   });
 });
