+++ e2e-tests/insurance/cypress/e2e/journeys/policy/check-your-answers/summary-list/single-policy/check-your-answers-single-other-company-no.spec.js
@@ -1,0 +1,52 @@
+import checkSummaryList from '../../../../../../../../commands/insurance/check-policy-summary-list';
+import { POLICY as POLICY_FIELD_IDS } from '../../../../../../../../constants/field-ids/insurance/policy';
+import { INSURANCE_ROUTES } from '../../../../../../../../constants/routes/insurance';
+
+const { ROOT: INSURANCE_ROOT, POLICY } = INSURANCE_ROUTES;
+
+const {
+  REQUESTED_JOINTLY_INSURED_PARTY: { REQUESTED, COMPANY_NAME, COMPANY_NUMBER, COUNTRY_CODE },
+} = POLICY_FIELD_IDS;
+
+const baseUrl = Cypress.config('baseUrl');
+
+context('Insurance - Policy - Check your answers - Summary list - Single contract policy - Other company as no', () => {
+  let referenceNumber;
+  let url;
+
+  before(() => {
+    cy.completeSignInAndGoToApplication({}).then(({ referenceNumber: refNumber }) => {
+      referenceNumber = refNumber;
+
+      cy.completePolicySection({ sameName: false });
+
+      url = `${baseUrl}${INSURANCE_ROOT}/${referenceNumber}${POLICY.CHECK_YOUR_ANSWERS}`;
+    });
+  });
+
+  beforeEach(() => {
+    cy.saveSession();
+
+    cy.navigateToUrl(url);
+  });
+
+  after(() => {
+    cy.deleteApplication(referenceNumber);
+  });
+
+  it(`should render a ${REQUESTED} summary list row`, () => {
+    checkSummaryList.REQUESTED_JOINTLY_INSURED_PARTY[REQUESTED]({});
+  });
+
+  it(`should NOT render a ${COMPANY_NAME} summary list row`, () => {
+    checkSummaryList.REQUESTED_JOINTLY_INSURED_PARTY[COMPANY_NAME]({});
+  });
+
+  it(`should NOT render a ${COMPANY_NUMBER} summary list row`, () => {
+    checkSummaryList.REQUESTED_JOINTLY_INSURED_PARTY[COMPANY_NUMBER]({});
+  });
+
+  it(`should NOT render a ${COUNTRY_CODE} summary list row`, () => {
+    checkSummaryList.REQUESTED_JOINTLY_INSURED_PARTY[COUNTRY_CODE]({});
+  });
+});
